A new file was created where all the code that creates the view was placed, which leaves the ViewControler much cleaner.
